C2T configuration for initiating the parameter values in the study environment

C2T configuration for initiating the parameter values in the study environment

Usage

cohort2TrajectoryConfiguration(
  cdm = NULL,
  studyName = "Cohort2Trajectory",
  baseUrl = "http://localhost:8080/WebAPI",
  pathToStudy = getwd(),
  atlasTargetCohort = NULL,
  atlasStateCohorts = NULL,
  stateCohortLabels = NULL,
  stateCohortPriorityOrder = NULL,
  stateCohortMandatory = NULL,
  stateCohortAbsorbing = NULL,
  stateSelectionType = NULL,
  oocFix = "None",
  trajectoryType = "Discrete",
  lengthOfStay = NULL,
  outOfCohortAllowed = FALSE,
  runSavedStudy = FALSE,
  useCDM = TRUE,
  trajectoryDataObject = NULL,
  pathToData = "./Cohort2Trajectory/Data/importedData.csv",
  allowedStatesList = createStateList(stateCohortLabels),
  mergeStates = FALSE,
  mergeThreshold = 0.5,
  batchSize = 1000
)

Arguments

cdm

object created with CDMConnector

studyName

name chosen for study, is used as a folder name

baseUrl

the base URL for the WebApi instance

pathToStudy

path to folder which contains folder named after studyName

atlasTargetCohort

the id of the target cohort defined in OHDSI tool ATLAS

atlasStateCohorts

the ids of the state cohorts defined in OHDSI tool ATLAS

stateCohortLabels

vector of the customized labels of the state cohorts

stateCohortPriorityOrder

vector of the customized labels of the state cohorts in priority order

stateCohortMandatory

vector of the customized labels of the state cohorts which are mandatory in trajectory

stateCohortAbsorbing

vector of the customized labels of the state cohorts which are absorbing

stateSelectionType

the type of state selection ("First" - First occurring, "Overlap" - Max overlap, "Priority" - Priority)

oocFix

the method to use for replacing "OUT OF COHORT" states with more relevant states ("None" -> "OUT OF COHORT"; "Last present state" -> repeat the last one; random str -> used as state)

trajectoryType

The type of the trajectory ("Discrete" - Discrete time, "Continuous" - Continuous time)

lengthOfStay

The length of stay (days) in one state (Effect only in discrete case)

outOfCohortAllowed

boolean whether the patient trajectory can surpass the target cohort's observation-period

runSavedStudy

running a predefined study from studyName/Settings/trajectorySettings.csv

useCDM

The package can also be run without the OMOP CDM

trajectoryDataObject

When using without OMOP CDM specify the data file (if specified no need for pathToData)

pathToData

When using without OMOP CDM specify the path to data file (if specified no need for trajectoryDataObject)

allowedStatesList

A list object which indicates accessible states from said state

mergeStates

Boolean, if you want to merge states when they overlap

mergeThreshold

Value from 0 to 1. If mergeStates is TRUE the states will be label-merged given they overlap more than the specified threshold. Can be given as vector, then multiple iterations are runned,

batchSize

customizable batch size for cohort generation

Examples

if (FALSE) cohort2TrajectoryConfiguration(cdm = cdm,
studyName = "TestCohort2Trajectory",
baseUrl = NULL,
pathToStudy = getwd(),
atlasTargetCohort = 0,
atlasStateCohorts = c(1, 2),
stateCohortLabels = c("test_label1", "test_label2"),
) # \dontrun{}