Import the data from CDM for the study

Import the data from CDM for the study

Usage

getDataForStudy(
  runSavedStudy = studyEnv$runSavedStudy,
  useCDM = studyEnv$useCDM,
  trajectoryDataObject = NULL,
  cdm = NULL,
  studyEnv = NULL,
  atlasStateCohorts = studyEnv$atlasStateCohorts,
  atlasTargetCohort = studyEnv$atlasTargetCohort,
  stateCohortLabels = studyEnv$stateCohortLabels,
  baseUrl = studyEnv$baseUrl,
  outOfCohortAllowed = studyEnv$outOfCohortAllowed,
  stateCohortMandatory = studyEnv$stateCohortMandatory,
  stateCohortPriorityOrder = studyEnv$stateCohortPriorityOrder,
  mergeStates = studyEnv$mergeStates,
  mergeThreshold = studyEnv$mergeThreshold,
  pathToData = NULL
)

Arguments

runSavedStudy

running a predefined study from studyName/Settings/trajectorySettings.csv

useCDM

The package can also be run without the OMOP CDM

trajectoryDataObject

When using without OMOP CDM specify the data file (if specified no need for pathToData)

cdm

object created with CDMConnector

studyEnv

environment created with cohort2TrajectoryConfiguration

atlasStateCohorts

the ids of the state cohorts defined in OHDSI tool ATLAS

atlasTargetCohort

the id of the target cohort defined in OHDSI tool ATLAS

stateCohortLabels

vector of the customized labels of the state cohorts

baseUrl

the base URL for the WebApi instance

outOfCohortAllowed

boolean whether the patient trajectory can surpass the target cohort's observation-period

stateCohortMandatory

vector of the customized labels of the state cohorts which are mandatory in trajectory

stateCohortPriorityOrder

vector of the customized labels of the state cohorts in priority order

mergeStates

Boolean, if you want to merge states when they overlap

mergeThreshold

Value from 0 to 1. If mergeStates is TRUE the states will be label-merged given they overlap more than the specified threshold. Can be given as vector, then multiple iterations are runned,

pathToData

When using without OMOP CDM specify the path to data file (if specified no need for trajectoryDataObject)

Value

log info

Examples

if (FALSE) getDataForStudy(cdm = cdm,studyEnv = studyEnv) # \dontrun{}