Prepare patientData dataframe for msm package

This function prepares patientData dataframe for msm package

Usage

getTrajectoriesContinuous(
  cdm,
  patientData,
  stateSelection,
  statePriorityVector,
  absorbingStates = NULL,
  studyName = "",
  addPersonalData = TRUE,
  allowedStatesList = list()
)

Arguments

cdm

a cdm object

patientData

Object of class data.frame with columns subject_id, state_label, STATE_START_DATE, STATE_END_DATE

stateSelection

Selection of the type of ordering

statePriorityVector

All states in prioritized order

absorbingStates

Absorbing states in dataframe

studyName

name chosen for study, is used as a folder name

addPersonalData

Logical, indicating whether or not to add personal data for each patient

allowedStatesList

A list object which indicates accessible states from said state

Value

A dataframe ready for using msm package